Martin's monster merge part 5/N: final cleanups, fixes and new testcases

See merge request !68

+++ nifty/spaces/power_space/power_space.py
@@ -119,7 +119,7 @@ class PowerSpace(Space):
 
     def get_distance_array(self, distribution_strategy):
         result = d2o.distributed_data_object(
-                                self.kindex,
+                                self.kindex, dtype=np.float64,
                                 distribution_strategy=distribution_strategy)
         return result
